% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_a_centrigrados.R
\name{fahrenheit_a_centigrados}
\alias{fahrenheit_a_centigrados}
\title{Función para convertir Fahrenheit a centigrados}
\usage{
fahrenheit_a_centigrados(temperatura_f)
}
\arguments{
\item{temperatura_f}{temperatura en fahrenheit}
}
\value{
Esta funcion devuelve la temperatura en centigrados
}
\description{
Esta funcion sirve para pasar temperaturas de Fahrenheit a centigrados
}
\examples{
fahrenheit_a_centigrados(98.6)

}
