% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion_grafico.R
\name{grafico_temperatura_mensual}
\alias{grafico_temperatura_mensual}
\title{Grafico de temperatura mensual}
\usage{
grafico_temperatura_mensual(
  estaciones,
  colores,
  titulo = "Temperatura mensual"
)
}
\arguments{
\item{estaciones}{Data frame o tibble con los datos de las estaciones. Debe contener las columnas: id, fecha, temperatura_abrigo_150cm.}

\item{colores}{Vector de colores para las líneas del gráfico. Si no se proporcionan suficientes colores, se generarán colores aleatorios.}

\item{titulo}{Título del gráfico. Por defecto es "Temperatura mensual".}
}
\value{
Un gráfico de líneas que representa la temperatura mensual por id.
}
\description{
' Esta función genera un gráfico de líneas que muestra la temperatura promedio mensual para cada estación en el conjunto de datos proporcionado.
}
\examples{
grafico_temperatura_mensual(estacion_NH0910, colores = c("red", "blue"), titulo = "Temperatura mensual NH0910")
}
