% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lee_archivos.R
\name{leer_archivo}
\alias{leer_archivo}
\title{Leer archivos}
\usage{
leer_archivo(ruta)
}
\arguments{
\item{ruta}{}
}
\value{
Te devuelve el archivo leido
}
\description{
Lo que hace esta funcion es leer tu archivo,
puede ser tanto URL como rutas locales
}
\details{
# lee archivos
}
