% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion_datos.R
\name{leer_datos_estacion}
\alias{leer_datos_estacion}
\title{Funcion Para Leer O Descargar Datos De Una Estacion Meteorologica}
\usage{
leer_datos_estacion(id_estacion, ruta_archivo)
}
\arguments{
\item{id_estacion}{ID de la estacion meteorologica. Valores validos: metadatos, NH0472, NH0910, NH0046, NH0098, NH0437}

\item{ruta_archivo}{Ruta local donde se encuentra o se descargara el archivo CSV con los datos de la estacion.}
}
\value{
Un data frame con los datos de la estacion meteorologica.
}
\description{
Esta funcion permite leer los datos si el archivo ya se encuentra cargado o cargarlo primero y luego leerlo si este no se encuentra.
}
\examples{
# leer_datos_estacion("NH0472", "NH0472.csv")
}
