% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_por_mes.R
\name{min_max_por_mes}
\alias{min_max_por_mes}
\title{Minimos y maximos por por fecha}
\usage{
min_max_por_mes(datos)
}
\value{
Poner el mes que deseamos ver
Ejemplo:
}
\description{
La funcion 'min_max_por_mes()' te devuelve las temperaturas minimas y maximas registradas en cada mes
}
\details{
Esta es informacion extra
}
\examples{
min_max_por_mes()
}
