% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_meto.R
\name{simular_meteorologia}
\alias{simular_meteorologia}
\title{Simular meteorologia}
\usage{
simular_meteorologia()
}
\value{
Dos valores, uno haciendo referencia a la temperatura, y otro a la presipitacion
Ejemplo: 10, 38
}
\description{
La funcion \code{simular_meteorologia()} te devuelve de manera random una temperatura y una precipitacion, no hace uso de parametros
}
\details{
Esta es informacion extra
}
\examples{
simular_meteorologia()

}
