% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion_resumen.R
\name{tabla_resumen_temperatura}
\alias{tabla_resumen_temperatura}
\title{Tabla resumen de temperatura por estación}
\usage{
tabla_resumen_temperatura(estacion)
}
\arguments{
\item{estacion}{Data frame o tibble con los datos de las estaciones. Debe contener las columnas: id, temperatura_abrigo_150cm, temperatura_abrigo_150cm_maxima, temperatura_abrigo_150cm_minima.}
}
\value{
Un data frame con las siguientes columnas: id, promedio_temperatura, desvio_estandar, temp_max, temp_min.
}
\description{
Esta función genera una tabla resumen con estadísticas de temperatura para cada estación. Incluye el promedio, desvío estándar, máximo y mínimo de temperatura
}
\examples{
tabla_resumen_temperatura(estacion_NH0910)
}
