% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_temperatura_mensual.R
\name{grafico_temperatura_mensual}
\alias{grafico_temperatura_mensual}
\title{Función de Gráfico de la Temperatura Mensual}
\usage{
grafico_temperatura_mensual(df, colores = NULL, titulo = "Temperatura")
}
\arguments{
\item{df, }{data frame con información de estaciones}

\item{colores, }{string o vector de colores}

\item{titulo}{string}
}
\value{
Gráfico de ggplot con la temperatura mensual de cada estación
}
\description{
La función se encarga de generar un gráfico de lineas para, a través de sus parámetros,
mostrar la temperatura promedio mensual de cada estación. Desde los datos, caclula el promedio
de temperatura mensual, y genera un gráfico comparando cada resultado.
}
\examples{
estaciones <- data.frame(
  id = c("NH0910", "NH0910"),
  fecha = as.Date(c("2020-01-01", "2020-02-01")),
  temperatura_abrigo_150cm = c(10, 12)
)
grafico_temperatura_mensual(estaciones, colores = c("red"), titulo = "Ejemplo"
}
