% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_datos_estacion.R
\name{leer_datos_estacion}
\alias{leer_datos_estacion}
\title{Función de lectura de los datos de las estaciones}
\usage{
leer_datos_estacion(id_estacion, ruta_archivo)
}
\arguments{
\item{id_estacion}{es un string del id}

\item{ruta_archivo}{string con la ruta del archivo, donde está guardado, o donde se quiera descargar.}
}
\value{
conjunto de datos de la estación/es llamadas
}
\description{
La función se encarga de leer los datos de los archivos de las estaciones, el usuario debe pasar los dos argumentos, siendo estos el id de la estación y la ruta para acceder a ellos, a su lectura y descarga.
}
\examples{
leer_datos_estacion("NHO472", "Downloads/Estaciones")

}
