% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_nubes.R, R/min_max_por_mes.R
\name{min_max_por_mes}
\alias{min_max_por_mes}
\title{Minimos y maximos por por fecha}
\usage{
min_max_por_mes(datos)

min_max_por_mes(datos)
}
\value{
Poner la estacion que deseamos ver
Ejemplo:

Poner el mes que deseamos ver
Ejemplo:
}
\description{
La funcion 'min_max_por_estacion()' te devuelve las temperaturas minimas y maximas registradas en cada mes

La funcion 'min_max_por_mes()' te devuelve las temperaturas minimas y maximas registradas en cada mes
}
\details{
Esta es informacion extra

Esta es informacion extra
}
\examples{
min_max_por_estacion()
min_max_por_mes()
}
