% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{simular_meteorologia}
\alias{simular_meteorologia}
\title{Simular meteorologia}
\usage{
simular_meteorologia()
}
\value{
10, 38
}
\description{
Lo que hace esta funcion, es devolverte de manera random una temperatura y una precipitacion, no hace uso de parametros
}
\details{
Esta es informacion extra
}
\examples{
simular_meteorologia()
}
